# persist: avoid overwriting meta in storage with exactly the same bytes