This introduces a performance and write amplification optimization
that avoids writing out to META if what we're writing matches what's
already in storage.

    indexed_write_drain/file_sorted/23000000
                            time:   [171.36 ms 172.77 ms 173.90 ms]
                            thrpt:  [126.14 MiB/s 126.96 MiB/s 128.00 MiB/s]
                     change:
                            time:   [-23.877% -22.099% -20.431%] (p = 0.00 < 0.05)
                            thrpt:  [+25.677% +28.368% +31.367%]
                            Performance has improved.

    indexed_write_drain/file_unsorted/23000000
                            time:   [185.01 ms 190.43 ms 197.72 ms]
                            thrpt:  [110.94 MiB/s 115.18 MiB/s 118.56 MiB/s]
                     change:
                            time:   [-22.737% -19.014% -15.377%] (p = 0.00 < 0.05)
                            thrpt:  [+18.171% +23.478% +29.428%]
                            Performance has improved.

### Motivation

  * This PR adds a feature that has not yet been specified.

### Tips for reviewer

I ran the background persisted system tables test with and without this change. Initially I didn't see much of a difference, but Ruchir reminded me in the PR comments that the need_step logic we have in RuntimeImpl is a workaround for exactly this issue. With the following diff applied (which more simply runs drain_pending and drain_unsealed each time though the runtime loop, but leaves compaction behind need_step since it's still overaggressive without it) we see much different results. After a few minutes of running the test:

Before this commit:
mz_persist_blob_meta_write_bytes 363620704
mz_server_metadata_seconds{...} 701.15673625
363620704 bytes / 701.15673625 s = 518 kBps

With this commit:
mz_persist_blob_meta_write_bytes 108652344
mz_server_metadata_seconds{...} 796.781403208
108652344 bytes / 796.781403208 s = 136 kBps

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly considered.
- [N/A] This PR adds a release note for any
  [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/user/content/release-notes.md#what-changes-require-a-release-note).