@ruchirK 2021-11-12
Also, I'm a little bit surprised by the improvement in the benchmarks because the benchmarks just call `write` and then `drain_pending` in a loop -- so it seems like the `try_set_meta` call inside `drain_pending` should always have real work to do (because we minted new unsealed batches), so it shouldn't be hitting this optimization?

wdyt am i missing something?