@danhhz 2021-11-12
> re: it not showing up as a perf improvement in mz -- it possible that its being covered up by the need_step optimization in runtime?