@danhhz 2021-11-12
Okay sanity checked my benchmark results by closing a bunch of open stuff on my laptop and running them again. Got similar results.

Looked at the code and it seems that the indexed_* benchmarks are calling `write` within `block_on_drain`, which internally calls `step` (not just a bare`drain_pending`). Given that the writes are never sealed, the `drain_unsealed` and `compact` parts of `step` have nothing to do. I _think_ we were doing the right thing in `compact` before this PR, but `drain_unsealed` is certainly becoming a no-op on each of those calls now, whereas before it was indeed overwriting an identical META.

Also, anecdotally, if I apply the following diff, then `cargo test -p persist --bench writer -- indexed` hits the panic and crashes out.

```diff
diff --git a/src/persist/src/indexed/mod.rs b/src/persist/src/indexed/mod.rs
index 81fe4a18e..b3ca112c4 100644
--- a/src/persist/src/indexed/mod.rs
+++ b/src/persist/src/indexed/mod.rs
@@ -269,7 +269,7 @@ impl<L: Log, B: Blob> Indexed<L, B> {
             // those checks would be difficult to maintain (and bugs in them
             // would surface as either unnecessary storage usage or correctness
             // issues).
-            return Ok(());
+            panic!("boom");
         }
         if let Err(e) = self.blob.set_meta(&new_meta) {
             // We were unable to durably commit the in-memory state. Revert back to the
```