@ruchirK 2021-11-12
That makes total sense, thanks for double checking! I forgot that the benchmarks call `step`.